fix(cli): extract avatar logic to fix complexity lint and validate before api calls
Move shared avatar resolution code into avatar.ts module to bring
resolveAvatarUrl complexity under the 20-branch limit. Also run
resolveAvatarUrl before getZeroAgent in applyAgentUpdate so validation
errors are thrown before any network request is made.
